+++ src/main/java/com/azure/cosmos/kafka/connect/CosmosDBConfig.java
@@ -1,14 +1,7 @@
 package com.azure.cosmos.kafka.connect;
 
-import static com.azure.cosmos.kafka.connect.CosmosDBConfig.CosmosClientBuilder.createClient;
-
 import com.azure.cosmos.CosmosClient;
 import com.azure.cosmos.kafka.connect.sink.CosmosDBSinkConfig;
-import java.net.InetAddress;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.UnknownHostException;
-import java.util.regex.Pattern;
 import org.apache.kafka.common.config.AbstractConfig;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.common.config.ConfigDef.Importance;
@@ -16,10 +9,12 @@ import org.apache.kafka.common.config.ConfigDef.NonEmptyString;
 import org.apache.kafka.common.config.ConfigDef.Type;
 import org.apache.kafka.common.config.ConfigDef.Validator;
 import org.apache.kafka.common.config.ConfigDef.Width;
+import org.apache.kafka.common.config.ConfigValue;
 
 import java.util.Map;
-import org.apache.kafka.common.config.ConfigException;
-import org.apache.kafka.common.config.ConfigValue;
+import java.util.regex.Pattern;
+
+import static com.azure.cosmos.kafka.connect.CosmosDBConfig.CosmosClientBuilder.createClient;
 
 @SuppressWarnings({"squid:S1854", "squid:S2160"})  // suppress unneeded int *groupOrder variables, equals method
 public class CosmosDBConfig extends AbstractConfig {
@@ -50,7 +45,21 @@ public class CosmosDBConfig extends AbstractConfig {
     private static final String COSMOS_CONTAINER_TOPIC_MAP_DISPLAY = "Topic-Container map";
     private static final String COSMOS_CONTAINER_TOPIC_MAP_DOC =
             "A comma delimited list of Kafka topics mapped to Cosmos containers.\n"
-                    + "For example: topic1#con1,topic2#con2.";    
+                    + "For example: topic1#con1,topic2#con2.";
+
+    public static final String COSMOS_CLIENT_TELEMETRY_ENABLED_CONF = "connect.cosmos.clientTelemetry.enabled";
+    public static final String COSMOS_CLIENT_TELEMETRY_ENABLED_DISPLAY = "Cosmos client telemetry enabled flag";
+    private static final String COSMOS_CLIENT_TELEMETRY_ENABLED_DOC =
+        "Cosmos client telemetry enabled flag";
+
+    public static final String COSMOS_CLIENT_TELEMETRY_ENDPOINT_CONF = "connect.cosmos.clientTelemetry.endpoint";
+    public static final String COSMOS_CLIENT_TELEMETRY_ENDPOINT_DISPLAY = "Cosmos client telemetry endpoint";
+    private static final String COSMOS_CLIENT_TELEMETRY_ENDPOINT_DOC =
+        "Cosmos client telemetry endpoint";
+
+    public static final String COSMOS_CLIENT_TELEMETRY_SCHEDULING_IN_SECONDS_CONF = "connect.cosmos.clientTelemetry.schedulingInSeconds";
+    public static final String COSMOS_CLIENT_TELEMETRY_SCHEDULING_IN_SECONDS_DISPLAY = "Cosmos client telemetry scheduling in seconds";
+    private static final String COSMOS_CLIENT_TELEMETRY_SCHEDULING_IN_SECONDS_DOC = "Cosmos client telemetry scheduling in seconds";
         
     public static final String COSMOS_PROVIDER_NAME_CONF = "connect.cosmos.provider.name";
     private static final String COSMOS_PROVIDER_NAME_DEFAULT = null;
@@ -67,6 +76,9 @@ public class CosmosDBConfig extends AbstractConfig {
     private String connKey;
     private String databaseName;
     private String providerName;
+    private boolean clientTelemetryEnabled;
+    private String clientTelemetryEndpoint;
+    private int clientTelemetrySchedulingInSeconds;
     private TopicContainerMap topicContainerMap = TopicContainerMap.empty();
 
     public CosmosDBConfig(ConfigDef config, Map<String, String> parsedConfig) {
@@ -77,6 +89,9 @@ public class CosmosDBConfig extends AbstractConfig {
         databaseName = this.getString(COSMOS_DATABASE_NAME_CONF);
         topicContainerMap = TopicContainerMap.deserialize(this.getString(COSMOS_CONTAINER_TOPIC_MAP_CONF));
         providerName = this.getString(COSMOS_PROVIDER_NAME_CONF);
+        clientTelemetryEnabled = this.getBoolean(COSMOS_CLIENT_TELEMETRY_ENABLED_CONF);
+        clientTelemetryEndpoint = this.getString(COSMOS_CLIENT_TELEMETRY_ENDPOINT_CONF);
+        clientTelemetrySchedulingInSeconds = this.getInt(COSMOS_CLIENT_TELEMETRY_SCHEDULING_IN_SECONDS_CONF);
     }
 
     public CosmosDBConfig(Map<String, String> parsedConfig) {
@@ -125,6 +140,27 @@ public class CosmosDBConfig extends AbstractConfig {
                         "none",
                         Importance.MEDIUM,
                         TOLERANCE_ON_ERROR_DOC
+                )
+                .define(
+                    COSMOS_CLIENT_TELEMETRY_ENABLED_CONF,
+                    Type.BOOLEAN,
+                    false,
+                    Importance.LOW,
+                    COSMOS_CLIENT_TELEMETRY_ENABLED_DOC
+                )
+                .define(
+                    COSMOS_CLIENT_TELEMETRY_ENDPOINT_CONF,
+                    Type.STRING,
+                    "https://tools.cosmos.azure.com/api/clienttelemetry/trace",
+                    Importance.LOW,
+                    COSMOS_CLIENT_TELEMETRY_ENDPOINT_DOC
+                )
+                .define(
+                    COSMOS_CLIENT_TELEMETRY_SCHEDULING_IN_SECONDS_CONF,
+                    Type.INT,
+                    600,
+                    Importance.LOW,
+                    COSMOS_CLIENT_TELEMETRY_SCHEDULING_IN_SECONDS_DOC
                 )
                 .defineInternal(
                         COSMOS_PROVIDER_NAME_CONF,
@@ -186,6 +222,30 @@ public class CosmosDBConfig extends AbstractConfig {
 
     public String getProviderName() {
         return this.providerName;
+    }
+
+    public boolean isClientTelemetryEnabled() {
+        return clientTelemetryEnabled;
+    }
+
+    public void setClientTelemetryEnabled(boolean clientTelemetryEnabled) {
+        this.clientTelemetryEnabled = clientTelemetryEnabled;
+    }
+
+    public String getClientTelemetryEndpoint() {
+        return clientTelemetryEndpoint;
+    }
+
+    public void setClientTelemetryEndpoint(String clientTelemetryEndpoint) {
+        this.clientTelemetryEndpoint = clientTelemetryEndpoint;
+    }
+
+    public int getClientTelemetrySchedulingInSeconds() {
+        return clientTelemetrySchedulingInSeconds;
+    }
+
+    public void setClientTelemetrySchedulingInSeconds(int clientTelemetrySchedulingInSeconds) {
+        this.clientTelemetrySchedulingInSeconds = clientTelemetrySchedulingInSeconds;
     }
 
     public static void validateConnection(Map<String, String> connectorConfigs, Map<String, ConfigValue> configValues) {